+++ Main/Features/NazgulFamily/NazgulFamilyIoC.cs
@@ -1,0 +1,11 @@
+using DryIoc;
+
+namespace TAOM.Features.NazgulFamily;
+
+public static class NazgulFamilyIoC
+{
+    public static void RegisterNazgulFamilyFeature(IContainer container)
+    {
+        container.Register<INazgulRegistry, NazgulRegistry>(Reuse.Singleton);
+    }
+}